refactor: rename settings feature to preferences edit
