+++ src/features/preferences-edit/ui/components/SettingsSection.spec.tsx
