+++ src/features/preferences-edit/ui/components/SettingsSection.tsx
