+++ src/pages/settings/ui/SettingsPage.tsx
@@ -4,7 +4,7 @@ import { useKey } from "react-use";
 
 import { useAuthAccount, useAuthUid } from "@/entities/auth";
 import { updatePreferences, usePreferences } from "@/entities/preferences";
-import { SettingsForm, usePreferencesFormState } from "@/features/settings";
+import { SettingsForm, usePreferencesFormState } from "@/features/preferences-edit";
 import { useSignIn } from "@/features/sign-in";
 import { useSignOut } from "@/features/sign-out";
 import { RemoteMutationNotice } from "@/shared/ui/remote-mutation-notice";
